refactoring: badge component  - component  - docs
