--- docs/components/badge/examples/BadgeLinksExample.vue
@@ -1,9 +1,0 @@
-<template>
-  <div class="vp-raw flex items-end">
-    <Badge href="#">Link</Badge>
-    <Badge size="sm" href="#">Link</Badge>
-  </div>
-</template>
-<script setup>
-import { Badge } from '../../../../src/index'
-</script>
