--- docs/components/badge/examples/BadgeSizesExample.vue
@@ -1,15 +1,0 @@
-<template>
-  <div class="vp-raw flex">
-    <Badge size="sm">Default</Badge>
-    <Badge type="dark" size="sm">Dark</Badge>
-    <Badge type="red" size="sm">Red</Badge>
-    <Badge type="green" size="sm">Green</Badge>
-    <Badge type="yellow" size="sm">Yellow</Badge>
-    <Badge type="indigo" size="sm">Indigo</Badge>
-    <Badge type="purple" size="sm">Purple</Badge>
-    <Badge type="pink" size="sm">Pink</Badge>
-  </div>
-</template>
-<script setup>
-import { Badge } from '../../../../src/index'
-</script>
