--- docs/components/badge/examples/BadgeTypesExample.vue
@@ -1,15 +1,0 @@
-<template>
-  <div class="vp-raw flex">
-    <Badge>Default</Badge>
-    <Badge type="dark">Dark</Badge>
-    <Badge type="red">Red</Badge>
-    <Badge type="green">Green</Badge>
-    <Badge type="yellow">Yellow</Badge>
-    <Badge type="indigo">Indigo</Badge>
-    <Badge type="purple">Purple</Badge>
-    <Badge type="pink">Pink</Badge>
-  </div>
-</template>
-<script setup>
-import { Badge } from '../../../../src/index'
-</script>
