+++ docs/components/badge/examples/FwbBadgeExample.vue
@@ -1,0 +1,32 @@
+<template>
+  <div class="vp-raw flex">
+    <fwb-badge>
+      Default
+    </fwb-badge>
+    <fwb-badge type="dark">
+      Dark
+    </fwb-badge>
+    <fwb-badge type="red">
+      Red
+    </fwb-badge>
+    <fwb-badge type="green">
+      Green
+    </fwb-badge>
+    <fwb-badge type="yellow">
+      Yellow
+    </fwb-badge>
+    <fwb-badge type="indigo">
+      Indigo
+    </fwb-badge>
+    <fwb-badge type="purple">
+      Purple
+    </fwb-badge>
+    <fwb-badge type="pink">
+      Pink
+    </fwb-badge>
+  </div>
+</template>
+
+<script setup>
+import { FwbBadge } from '../../../../src/index'
+</script>
