+++ docs/components/badge/examples/FwbBadgeExampleLink.vue
@@ -1,0 +1,17 @@
+<template>
+  <div class="vp-raw flex items-end">
+    <fwb-badge href="#">
+      Link
+    </fwb-badge>
+    <fwb-badge
+      href="#"
+      size="sm"
+    >
+      Link small
+    </fwb-badge>
+  </div>
+</template>
+
+<script setup>
+import { FwbBadge } from '../../../../src/index'
+</script>
